requirements.txt
gradio
pyannote.audio
pydub
torch
transformers
tqdm
moviepy
srt
ffmpeg-python
TTS
openai-whisper
pytube